Change in gradle for app release
