+++ app/src/main/java/com/aasc/solvecase/SecondYearSubjects.kt
@@ -1,0 +1,13 @@
+package com.aasc.solvecase
+
+import android.os.Bundle
+import androidx.appcompat.app.AppCompatActivity
+
+class SecondYearSubjects :AppCompatActivity(){
+    override fun onCreate(savedInstanceState: Bundle?) {
+        super.onCreate(savedInstanceState)
+
+        setContentView(R.layout.second_year_subjects)
+
+    }
+}
