+++ app/src/main/java/com/aasc/solvecase/SignupActivity.kt
@@ -1,0 +1,69 @@
+package com.aasc.solvecase
+
+import android.content.Intent
+
+import android.os.Bundle
+import android.util.Patterns
+import android.widget.Toast
+import androidx.appcompat.app.AppCompatActivity
+import com.google.firebase.auth.FirebaseAuth
+import kotlinx.android.synthetic.main.activity_signup.*
+
+class SignupActivity : AppCompatActivity() {
+
+    private lateinit var auth: FirebaseAuth
+
+    override fun onCreate(savedInstanceState: Bundle?) {
+        super.onCreate(savedInstanceState)
+        setContentView(R.layout.activity_signup)
+        auth = FirebaseAuth.getInstance()
+
+        btn_sign_up.setOnClickListener {
+            signUpUser()
+        }
+        back_btn.setOnClickListener {
+            val back_intent = Intent(this,LoginActivity::class.java)
+            startActivity(back_intent)
+        }
+
+    }
+
+    private fun signUpUser() {
+        if (tv_username.text.toString().isEmpty()) {
+            tv_username.error = "Please enter email"
+            tv_username.requestFocus()
+            return
+        }
+
+        if (!Patterns.EMAIL_ADDRESS.matcher(tv_username.text.toString()).matches()) {
+            tv_username.error = "Please enter valid email"
+            tv_username.requestFocus()
+            return
+        }
+
+        if (tv_password.text.toString().isEmpty()) {
+            tv_password.error = "Please enter password"
+            tv_password.requestFocus()
+            return
+        }
+
+        auth.createUserWithEmailAndPassword(tv_username.text.toString(), tv_password.text.toString())
+            .addOnCompleteListener(this) { task ->
+                if (task.isSuccessful) {
+                    val user = auth.currentUser
+                    user?.sendEmailVerification()
+                        ?.addOnCompleteListener { task ->
+                            if (task.isSuccessful) {
+                                startActivity(Intent(this, LoginActivity::class.java))
+                                finish()
+                            }
+                        }
+                } else {
+                    Toast.makeText(
+                        baseContext, "Sign Up failed. Try again after some time.",
+                        Toast.LENGTH_SHORT
+                    ).show()
+                }
+            }
+    }
+}
