--- app/src/main/java/com/example/jiitsolutions/SecondYearSubjects.kt
@@ -1,19 +1,0 @@
-package com.example.jiitsolutions
-
-import android.content.Intent
-import android.os.Bundle
-import android.view.View
-import android.widget.RadioButton
-import android.widget.Toast
-import androidx.appcompat.app.AppCompatActivity
-import kotlinx.android.synthetic.main.first_year_subjects.*
-import kotlinx.android.synthetic.main.second_year_subjects.*
-
-class SecondYearSubjects :AppCompatActivity(){
-    override fun onCreate(savedInstanceState: Bundle?) {
-        super.onCreate(savedInstanceState)
-
-        setContentView(R.layout.second_year_subjects)
-
-    }
-}
